backend/requirements.txt
cryptography
fastapi
openpyxl
passlib
psycopg2
python-jose
sqlalchemy
uvicorn

